pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0,<2"]
build-backend = "hatchling.build"

[project]
name = "lean-spec"
version = "0.0.1"
description = "Lean Ethereum protocol specifications"
readme = "README.md"
authors = [
    { name = "Ethereum Foundation", email = "thomas.coratger@ethereum.org" },
]
keywords = [
    "lean",
    "beam",
    "ethereum",
    "specifications",
    "protocol",
    "consensus",
]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.12.0,<3",
    "typing-extensions>=4.4",
    "lean-multisig-py>=0.1.0",
    "httpx>=0.28.0,<1",
    "aiohttp>=3.11.0,<4",
    "cryptography>=46.0.0",
    "numpy>=2.0.0,<3",
    "prometheus-client>=0.21.0,<1",
    "aioquic>=1.2.0,<2",
    "pyyaml>=6.0.0,<7",
]

[project.license]
file = "LICENSE"

[project.urls]
Homepage = "https://github.com/leanEthereum/leanSpec"
Source = "https://github.com/leanEthereum/leanSpec"
Issues = "https://github.com/leanEthereum/leanSpec/issues"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build]
exclude = [".*", "tests/", "docs/"]

[tool.hatch.build.targets.wheel]
packages = ["src/lean_spec"]

[tool.ruff]
line-length = 100

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["E", "F", "B", "W", "I", "A", "N", "D", "C"]
fixable = ["I", "B", "E", "F", "W", "D", "C"]
ignore = ["D205", "D203", "D212", "D415", "C901", "A005", "C420"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
force-single-line = false
known-first-party = ["lean_spec"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["D"]

[tool.ty.environment]
python-version = "3.12"

[tool.ty.terminal]
error-on-warning = true

[tool.pytest.ini_options]
minversion = "8.3.3"
testpaths = ["tests"]
python_files = "test_*.py"
pythonpath = ["."]
addopts = [
    "-ra",
    "--strict-markers",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-branch",
    # Exclude fixture generation tests from regular test runs
    # These are only run via the 'fill' command
    "--ignore=tests/consensus",
    "--ignore=tests/execution",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "valid_until: marks tests as valid until a specific fork version",
]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.mdformat]
number = true
wrap = 80

[tool.uv]
required-version = ">=0.7.0"

[tool.uv.workspace]
members = ["packages/*"]

[tool.uv.sources]
lean-ethereum-testing = { workspace = true }
lean-multisig-py = { git = "https://github.com/anshalshukla/leanMultisig-py", branch = "devnet2" }

[dependency-groups]
test = [
    "pytest>=8.3.3,<9",
    "pytest-cov>=6.0.0,<7",
    "pytest-xdist>=3.6.1,<4",
    "hypothesis>=6.138.14",
    "lean-ethereum-testing",
    "lean-multisig-py>=0.1.0",
]
lint = [
    "ty>=0.0.1a34",
    "ruff>=0.13.2,<1",
    "codespell>=2.4.1,<3",
]
docs = [
    "mkdocs>=1.6.1,<2",
    "mkdocs-material>=9.5.45,<10",
    "mkdocstrings[python]>=0.27.0,<1",
    "mdformat==0.7.22",
]
dev = [
    {include-group = "test"},
    {include-group = "lint"},
    {include-group = "docs"},
    "ipython>=8.31.0,<9",
    "ipdb>=0.13",
    "tomli-w>=1.0.0",
    "build>=1.2.0,<2",
    "twine>=5.1.0,<6",
]
